+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact agate module, newest cha
 - `agate/Panels` to show close button properly for night mode
 - `LabeledIconButton` max-width so that huge sized icon will not be cut off
 - `agate/Spinner` to pause the animation when `paused` prop is true
+- `agate/Spinner` changes color based on day/night mode and added `transparent` prop
 
 ## [1.0.0] - 2020-10-14